feature / доработан кастомный хук для отображения заданного количества результатов поиска в зависимости от типа устройства
