+++ src/components/MoviesCardList/MoviesCardList.js
@@ -2,20 +2,14 @@ import './MoviesCardList.css';
 
 import MoviesCard from "../MoviesCard/MoviesCard";
 
-export default function MoviesCardList({moviesArray, savedFilms, onAddBtnClick, addBtnVisible, currentMoviesAmount}) {
+export default function MoviesCardList({displayMovies, savedFilms, onAddBtnClick, addBtnVisible}) {
 
-  const displayMovies = [];
-
-  if ((moviesArray.length > 0) && (moviesArray.length >= currentMoviesAmount)) {
-    for (let i = 0; i < currentMoviesAmount; i++) {
-      displayMovies.push(<MoviesCard id={moviesArray[i].id} key={moviesArray[i].id} data={moviesArray[i]}
-                                     savedFilms={savedFilms}/>);
-    }
-  }
   return (
     <section className="films">
       <ul className="films__list">
-        {displayMovies}
+        {displayMovies.map(item => (
+          <MoviesCard id={item.id} key={item.id} data={item} savedFilms={savedFilms}/>
+        ))}
       </ul>
       {!savedFilms && addBtnVisible && <button className="films__button-more" onClick={onAddBtnClick}>Ещё</button>}
     </section>
